requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
django-taggit==3.0.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.4
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests-oauthlib==1.3.1
sqlparse==0.4.3
